Various updates to glob_dist.json
This file has gotten outdated compared to what's used in chef-workstation. In
addition, there are some fixes that needed to happen to help with Cinc:

- Sort the constants so it's easier to read (basically matches what
  chef-workstation has)
- Import constants from chef-workstation that haven't been added here
- Add InspecGem and ClientGem which will be used in the chef go CLI
  - Chef Workstation is currently relying on CLIWrapperExec and InspecCli which
    Cinc renames but the gem name is different/same
- Add InspecExec (which will replace InspecCli in chef-workstation) to match
  naming scheme

Once this has been merged, I will make a PR in chef-workstation to update the
global.go file and also update the code to use the new constants.

Signed-off-by: Lance Albertson <lance@osuosl.org>